// frontend/src/api/model/entryTypeEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `reset` - Reset
* `sync` - Sync
* `maintenance` - Maintenance
* `note` - Note
* `location_update` - Location update
* `calibration` - Calibration
* `other` - Other
 */
export type EntryTypeEnum = typeof EntryTypeEnum[keyof typeof EntryTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EntryTypeEnum = {
  reset: 'reset',
  sync: 'sync',
  maintenance: 'maintenance',
  note: 'note',
  location_update: 'location_update',
  calibration: 'calibration',
  other: 'other',
} as const;