// frontend/src/api/model/sourceEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * DOSPORTAL API
 * API Documentation for DOSPORTAL
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `web` - Web
* `api` - API
* `qr` - QR
* `auto` - Automatic
* `other` - Other
 */
export type SourceEnum = typeof SourceEnum[keyof typeof SourceEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SourceEnum = {
  web: 'web',
  api: 'api',
  qr: 'qr',
  auto: 'auto',
  other: 'other',
} as const;